resolve lint errors
